end of section file upload and download

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const Product = require('../models/product');
+const fileHelper = require('../util/file')
 const { validationResult } = require('express-validator/check');
 
 exports.getAddProduct = (req, res, next) => {
@@ -14,20 +15,36 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage: 'Attached files is not an image',
+      validationErrors: []
+    });
+  }
+
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
-      path: '/admin/edit-product',
+      path: '/admin/add-product',
       editing: false,
       hasError: true,
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description
       },
@@ -35,6 +52,8 @@ exports.postAddProduct = (req, res, next) => {
       validationErrors: errors.array()
     });
   }
+
+  const imageUrl = image.path;
 
   const product = new Product({
     title: title, 
@@ -52,6 +71,7 @@ exports.postAddProduct = (req, res, next) => {
     res.redirect('/admin/products')
   })
   .catch(err => {
+    // console.log(err)
     const error = new Error(err);
     error.httpStatusCode = 500;
     return next(error);
@@ -91,7 +111,7 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updateTitle = req.body.title;
-  const updateImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatePrice = req.body.price;
   const updateDesc = req.body.description;
 
@@ -105,7 +125,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updateTitle,
-        imageUrl: updateImageUrl,
         price: updatePrice,
         description: updateDesc,
         _id: prodId
@@ -120,10 +139,13 @@ exports.postEditProduct = (req, res, next) => {
       if (product.userId.toString() !== req.user._id.toString()) {
         return res.redirect('/')
       }
-      product.title = updateTitle,
-      product.imageUrl = updateImageUrl,
-      product.price = updatePrice,
-      product.description = updateDesc
+      product.title = updateTitle;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
+      product.price = updatePrice;
+      product.description = updateDesc;
       return product.save()
     })
     .then(() => {
@@ -138,10 +160,19 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId).then(product => {
+    if (!product) {
+      return next(new Error('No Product Found'))
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    return Product.deleteOne({ _id: prodId, userId: req.user._id })
+  })
   .then(() => res.redirect('/admin/products'))
-  .catch(err => console.log(err));
-  
+  .catch(err => {
+    const error = new Error(err);
+    error.httpStatusCode = 500;
+    return next(error);
+  });
 }
 
 exports.getProducts = (req, res, next) => {
